requirements.txt
blinker==1.7.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
dill==0.3.7
Flask==3.0.0
fonttools==4.42.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipywidgets==8.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jupyterlab-widgets==3.0.8
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
mysql-connector-python==8.2.0
pandas==2.0.3
pbr==5.11.1
Pillow==10.0.0
ply==3.11
protobuf==4.21.12
pydot==1.4.2
pylatexenc==2.10
pyparsing==3.0.9
pytz==2023.3.post1
rustworkx==0.13.1
scipy==1.10.1
stevedore==5.1.0
symengine==0.9.2
sympy==1.12
tzdata==2023.3
Werkzeug==3.0.1
widgetsnbextension==4.0.8
zipp==3.17.0
